+++ Model/PatternDatabase/PatternDatabase.h
@@ -35,11 +35,11 @@ namespace busybin
 
   public:
     PatternDatabase(const size_t size);
-    virtual uint32_t getDatabaseIndex(const RubiksCube& cube) const = 0;
+    virtual size_t getDatabaseIndex(const RubiksCube& cube) const = 0;
     virtual bool setNumMoves(const RubiksCube& cube, const uint8_t numMoves);
-    virtual bool setNumMoves(const uint32_t ind, const uint8_t numMoves);
+    virtual bool setNumMoves(const size_t ind, const uint8_t numMoves);
     virtual uint8_t getNumMoves(const RubiksCube& cube) const;
-    virtual uint8_t getNumMoves(const uint32_t ind) const;
+    virtual uint8_t getNumMoves(const size_t ind) const;
     virtual size_t getSize() const;
     virtual size_t getNumItems() const;
     virtual bool isFull() const;
